CLI: restructure for validators
