+++ cli/commands/preview/create/index.ts
@@ -3,11 +3,11 @@ import path from 'path';
 import { __, sprintf } from '@wordpress/i18n';
 import { getWordPressVersion } from 'common/lib/get-wordpress-version';
 import { PreviewCommandLoggerAction as LoggerAction } from 'common/logger-actions';
+import { siteSizeValidator } from 'cli/commands/preview/create/validator/site-size';
 import { uploadArchive, waitForSiteReady } from 'cli/lib/api';
 import { getAuthToken, getSiteByFolder } from 'cli/lib/appdata';
 import { archiveSiteContent, cleanup } from 'cli/lib/archive';
 import { saveSnapshotToAppdata } from 'cli/lib/snapshots';
-import { validateSiteSize } from 'cli/lib/validation';
 import { Logger, LoggerError } from 'cli/logger';
 import { StudioArgv } from 'cli/types';
 
@@ -21,7 +21,7 @@ export async function runCommand( siteFolder: string ): Promise< void > {
 	try {
 		logger.reportStart( LoggerAction.VALIDATE, __( 'Validating…' ) );
 		await getSiteByFolder( siteFolder );
-		await validateSiteSize( siteFolder );
+		await siteSizeValidator( siteFolder );
 		const token = await getAuthToken();
 		logger.reportSuccess( __( 'Validation successful' ), true );
